extract the zod validations to utils
